admin/requirements.txt
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
django-cors-headers==3.5
gunicorn==21.2.0
Pillow==10.0.0
pika==1.3.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python-dateutil==2.8.2
